initial commit
Started data cleaning process for CIR by replicating what was done for par. This resulted in significant loss of data. I'll need to go back and revise data cleaning strategy so there are CIRs for all the RFlabels in the plot.

Still working on figuring out how to link the mean differences dataset to corresponding RFlabels.